+++ test/e2e/buckets.sh
@@ -111,6 +111,7 @@ EOF
 
 bucket_api_mods() { cat <<'EOF'
 TestAPI_ModManifestInstallUpgrade
+TestAPI_ModUpload
 EOF
 }